+++ src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/fluid_shell/FluidShellBlockItem.java
@@ -4,14 +4,15 @@ import java.util.List;
 
 import javax.annotation.Nullable;
 
+import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
 import rbasamoyai.createbigcannons.multiloader.IndexPlatform;
 import rbasamoyai.createbigcannons.munitions.FuzedProjectileBlockItem;
 
@@ -30,11 +31,9 @@ public class FluidShellBlockItem extends FuzedProjectileBlockItem {
 	}
 
     @Override
-    public InteractionResult place(BlockPlaceContext context) {
-        InteractionResult result = super.place(context);
-        if (context.getLevel().getBlockEntity(context.getClickedPos()) instanceof AbstractFluidShellBlockEntity be)
-            be.readFluidDataFromFluidShellItem(context.getItemInHand(), context.getLevel().registryAccess());
-        return result;
+    protected boolean updateCustomBlockEntityTag(BlockPos pos, Level level, @org.jetbrains.annotations.Nullable Player player, ItemStack stack, BlockState state) {
+        if (level.getBlockEntity(pos) instanceof AbstractFluidShellBlockEntity be)
+            be.readFluidDataFromFluidShellItem(stack, level.registryAccess());
+        return super.updateCustomBlockEntityTag(pos, level, player, stack, state);
     }
-
 }