+++ src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/propellant/BigCartridgeBlockItem.java
@@ -6,11 +6,9 @@ import javax.annotation.Nullable;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
 import rbasamoyai.createbigcannons.CreateBigCannons;
 import rbasamoyai.createbigcannons.base.CBCTooltip;
@@ -34,19 +32,6 @@ public class BigCartridgeBlockItem extends BlockItem {
 		CBCTooltip.appendPropellantStressText(stack, level, tooltipComponents, isAdvanced, this.cartridgeBlock);
 		if (stack.getOrCreateTag().getBoolean("Damp"))
 			tooltipComponents.add(Component.translatable("block." + CreateBigCannons.MOD_ID + ".propellant.tooltip.damp").withStyle(ChatFormatting.BLUE));
-	}
-
-	@Override
-	public InteractionResult place(BlockPlaceContext context) {
-		InteractionResult result = super.place(context);
-
-		Level level = context.getLevel();
-		if (!level.isClientSide && level.getBlockEntity(context.getClickedPos()) instanceof BigCartridgeBlockEntity cart) {
-			cart.setPower(getPower(context.getItemInHand()));
-			cart.setChanged();
-		}
-
-		return result;
 	}
 
 	public int getMaximumPowerLevels() {